We are refactoring the feature related to CRUD operations on the folder entity. We decided to remove the AppRepository.kt and instead reference FolderDao.kt in FoldersViewModel.kt. In order to keep using the delegate property by viewModels(), we used the HiltViewModel attribute and also build a custom Hilt module, namely, DatabaseModule.kt.
On the other hand, we changed the UI for the creation of new folders. We no longer use the floating action button from MainActivity.kt, but instead we have a "header" item on the list of folders in FoldersFragment.kt which upon click on it displays the dialog in FolderDialogFragment.kt.