We made another big refactoring. We decided to create a custom data class FolderUiState to encapsulate the UI state when the user creates or updates a folder. On the other hand, we also made a sealed class InputNameState to specify the different states of the name of a folder when the user was typing on the AlertDialog. This allowed us to update the UI of the dialog accordingly based on which state the name inputted by the user was (see FolderState.kt and FolderDialogFragment.kt).
On the other hand, we decided to expose LiveData objects to the UI layer (i.e. activities and fragments) instead of flows, mainly because it's easier to observe a livedata than collect a flow. This allowed us to remove unnecessary extension methods for the Activity and Fragment classes (see UiUtils.kt).

Lastly, we update the Gradle plugin (see build.gradle, gradle-wrapper.properties and AndroidManifest.xml).