We decided to revamp the UI of the app. Instead of using a drawer layout together with a navigation view, we decided to use a bottom navigation view because we felt we didn't have enough items to show in the drawer. This change in the UI implied a major refactoring of several classes (see e.g., NotesFragment.kt, FoldersFragment.kt, EditorFragment.kt, etc.).
We also decided to show the floating action button on the start destination, namely, FoldersFragment.kt in order to allow users to create notes without too many clicks (by default notes created like that are located in the General folder).

Finally, we also added a couple of binding adapters for the empty view of a CollectionView and for the navigation icon click listener of the MaterialToolbar (see BindingAdapters.kt).