We rolled back the changes in the UI of the app. Ultimately, we felt that the bottom navigation UI wasn't the best fit for the app and it also didn't improve much the UX. By contrast, now users can see all their notes in one place directly. And yes, we removed the concept of folders and now the notes are grouped by notebooks (see NotebookEntity.kt, Notebook.kt, NotebookDao.kt and NotebookUseCases.kt).
On the other hand we upgrade several dependencies in build.gradle and that forced us to refactor the code due to nullability issues (see FullScreenImageFragment.kt).

Finally, we fixed the bug where the EditText to compose notes was losing focus after the soft keyboard was shown on the screen (see EditorContentAdapter.kt).