We finished the feature that empties the recycle bin and also deletes all the unused images saved in the internal storage of the app. This was made in order to save some space for the user. To accomplish such feature we had to made a synchronous version of the method to get the trashed notes that didn't required to return a Flow (see NoteDao.kt and NoteUseCases.kt). This values is then used when the RecycleBinViewModel.kt is instantiated (see also  EmptyRecycleBinDialogFragment.kt).
On the other hand, we also added the action mode for the RecycleBinFragment.kt with two actions: to delete permanently a note or to restore it. The first action is done completely, however, the second one is still in development (see RecyclerViewActionModeCallback.kt and recyclerview_trash_context_menu.xml).

Finally, to simplify code, we added a method in ImageStorageManager.kt to delete images using the ImageEntity data class, instead of just using the filename.