We finished the implementation of the feature that allows users to add a content description for a given image they attach to a note (see EditorFragment.kt, EditorViewModel.kt, AltTextState.kt, ImageAdapter.kt and ImageDescriptionDialogFragment.kt). We decided to allow the user to only add the alt text to images that are already saved.
On the other hand, we added a default content description in case images don't have any added by the user (see strings.xml, image_item.xml and single_image_item.xml).

Finally, we increase the width stroke of the MaterialCardView in note_item.xml and increased the min height of the EditText in fragment_editor_content.xml in order to fix an accessibility issue.