We finally completed the feature to restore notes that are sent to the recycle bin. In the situation where the notebook that contained the note(s) was also deleted, the note(s) will be only visible in the AllNotes view. See NotebookDao.kt, NoteDao.kt, NotebookUseCases.kt, NoteUseCases.kt, NotesViewModel.kt, NotesFragment.kt and RecycleBinViewModel.kt.
On the other hand, we decided to create two default notebooks: Personal and Work in order to simplify the grouping of notes for the user. See AppDatabaseCallback.kt and DatabaseModule.kt.

Lastly, we fixed the bug where after emptying the recycle bin, the table Images wasn't updated appropriately. See ImageDao.kt and NoteUseCases.kt.