We started the implementation of the feature that allows users to change the ordering of the notes. Currently three types of sorting are supported: by modification date, by content, and by title. See NotesFragment.kt, NotesMenuListener.kt, NotesMenuProvider.kt and SortNotesDialogFragment.kt.
On the other hand, we fixed the warning: "The following options were not recognized by any processor: '[room.schemaLocation, dagger.fastInit, kapt.kotlin.generated]'" thrown by AS. It was an innocuous warning, but annoying nonetheless. The fix was to place last the id kotlin-kapt inside the plugins sections in build.gradle.