We finished the implementation of the feature that allows users to pin notes. See NoteUseCases.kt, RecyclerViewActionModeCallback.kt, SelectionObserver.kt, NoteAdapter.kt, NotesFragment.kt and NotesViewModel.kt. We also had to change the layout of the protected notes (see protected_note_item.xml)
On the other hand, we added the right title to the icon for changing the layout of the notes (see NotesMenuProvider.kt). And finally, we fixed a bug where the intent to confirm the device credentials was called, even if there wasn't any credentials. This caused a NPE when trying to open a protected note. Now, if the user doesn't have any credentials, a Snackbar is shown informing the user about the lack of credentials (see NotesFragment.kt).