We realized that we don't need to get a reference to the note saved into the arguments of EditorFragment.kt when navigating from a note in NotesFragment.kt to EditorFragment.kt because that value is also stored into the saved state handle that we pass to EditorViewModel.kt. Hence, we only need that property from EditorViewModel.kt (see also NoteAdapter.kt). With this change, it was easier to create headless fragments to do operations that require a reference to a Context. And to avoid any complications with this kind of fragments, we ensured that they are children of EditorFragment.kt, and that they share the same instance of EditorViewModel.kt, and that when EditorFragment.kt is removed its children are also removed (see BaseHeadlessChildFragment.kt, AddImagesFragment.kt and ShareContentFragment.kt).
With these changes, it was easier to lock or pin notes while creating them, i.e., the user can lock or pin a note from within the Editor screen. This also made us ensure to have in sync the state of the menu items in the EditorFragment's toolbar (see fragment_editor.xml, menu_editor.xml, ToolbarItemMenuListener.kt and MenuItemClickListener.kt). Lastly, is important to remark that the changes made will simplify the implementation of the feature that allows to change the background color of the editor.